feat(storage): add text search capability and refactor example structure
- Add search_text() method to ChunkStore trait for substring search
- Implement SQL-based text search in SqliteStore with proper escaping
- Refactor end-to-end example to use include_str\!() for better maintainability
- Extract inline test data into structured test_data/ directory
- Support case-sensitive and case-insensitive text search queries
