+++ janet-ai-retriever/examples/test_data/README.md
@@ -1,0 +1,75 @@
+# Test Repository
+
+This is a test repository containing various programming examples to demonstrate the indexing and search capabilities of the janet-ai-retriever system.
+
+## Contents
+
+### Rust Library (`src/`)
+
+A sample Rust library demonstrating modern Rust programming patterns:
+
+- **lib.rs**: Main library module with configuration and initialization
+- **math.rs**: Mathematical utility functions including basic arithmetic and algorithms
+- **http.rs**: HTTP client utilities with request/response handling
+- **database.rs**: Database connection pooling and query execution
+- **auth.rs**: User authentication and authorization system
+
+### Python Utilities (`utils.py`)
+
+Data processing and analysis utilities written in Python:
+
+- Statistical calculations for numerical data
+- JSON file processing with error handling
+- Text formatting and transformation functions
+- Object-oriented data processing with the `DataProcessor` class
+
+### JavaScript API Client (`api.js`)
+
+A comprehensive HTTP client for RESTful API interactions:
+
+- Promise-based HTTP methods (GET, POST, PUT, DELETE)
+- Request and response interceptors
+- Authentication token management
+- Error handling and timeout support
+- Utility functions for common operations
+
+## Features Demonstrated
+
+The code in this repository showcases various programming concepts:
+
+1. **Error Handling**: Proper error types and handling patterns in Rust, Python, and JavaScript
+2. **Async Programming**: Async/await patterns in all three languages
+3. **Object-Oriented Design**: Classes and modules with clear separation of concerns
+4. **Documentation**: Comprehensive documentation with examples and type hints
+5. **Configuration Management**: Environment-based configuration and settings
+6. **Network Programming**: HTTP clients and server utilities
+7. **Data Processing**: Statistical analysis and data transformation
+8. **Authentication**: Token-based authentication systems
+9. **Database Operations**: Connection pooling and transaction management
+
+## Usage
+
+This repository is primarily used for testing the semantic search capabilities of the janet-ai-retriever system. The varied content allows for testing different types of queries:
+
+- Function definitions and implementations
+- Error handling patterns
+- Configuration and setup procedures
+- Mathematical operations and algorithms
+- Network and HTTP operations
+- Database interactions
+- Authentication workflows
+
+## Testing Queries
+
+Good test queries for this codebase include:
+
+- "function that adds two numbers"
+- "HTTP request handling"
+- "database connection management"
+- "user authentication system"
+- "error handling patterns"
+- "configuration setup"
+- "mathematical calculations"
+- "async programming examples"
+
+The semantic search should be able to find relevant code snippets for each of these concepts across the different programming languages and modules.
