+++ janet-ai-retriever/examples/test_data/src/database.rs
@@ -1,0 +1,143 @@
+//! Database connection and query utilities
+
+use std::collections::HashMap;
+use std::time::Duration;
+
+/// Database connection configuration
+#[derive(Debug, Clone)]
+pub struct DatabaseConfig {
+    pub host: String,
+    pub port: u16,
+    pub database: String,
+    pub username: String,
+    pub password: String,
+    pub max_connections: u32,
+    pub timeout: Duration,
+}
+
+/// Database connection pool
+pub struct DatabasePool {
+    config: DatabaseConfig,
+    active_connections: u32,
+}
+
+impl DatabasePool {
+    /// Create a new database pool
+    pub fn new(config: DatabaseConfig) -> Self {
+        Self {
+            config,
+            active_connections: 0,
+        }
+    }
+
+    /// Get a connection from the pool
+    pub async fn get_connection(&mut self) -> Result<DatabaseConnection, DatabaseError> {
+        if self.active_connections >= self.config.max_connections {
+            return Err(DatabaseError::PoolExhausted);
+        }
+
+        self.active_connections += 1;
+        Ok(DatabaseConnection {
+            id: self.active_connections,
+            config: self.config.clone(),
+        })
+    }
+
+    /// Return a connection to the pool
+    pub fn return_connection(&mut self, _connection: DatabaseConnection) {
+        self.active_connections = self.active_connections.saturating_sub(1);
+    }
+}
+
+/// Database connection
+pub struct DatabaseConnection {
+    id: u32,
+    config: DatabaseConfig,
+}
+
+impl DatabaseConnection {
+    /// Execute a SQL query and return results
+    pub async fn query(&self, sql: &str) -> Result<QueryResult, DatabaseError> {
+        println!("Executing query on connection {}: {}", self.id, sql);
+        
+        // Mock implementation
+        Ok(QueryResult {
+            rows: vec![],
+            affected_rows: 0,
+        })
+    }
+
+    /// Execute a prepared statement with parameters
+    pub async fn execute(&self, sql: &str, params: &[&str]) -> Result<QueryResult, DatabaseError> {
+        println!("Executing prepared statement on connection {}: {} with {} params", 
+                self.id, sql, params.len());
+        
+        // Mock implementation
+        Ok(QueryResult {
+            rows: vec![],
+            affected_rows: 1,
+        })
+    }
+
+    /// Begin a database transaction
+    pub async fn begin_transaction(&self) -> Result<DatabaseTransaction, DatabaseError> {
+        println!("Beginning transaction on connection {}", self.id);
+        Ok(DatabaseTransaction {
+            connection_id: self.id,
+        })
+    }
+}
+
+/// Database query result
+pub struct QueryResult {
+    pub rows: Vec<HashMap<String, String>>,
+    pub affected_rows: u64,
+}
+
+/// Database transaction
+pub struct DatabaseTransaction {
+    connection_id: u32,
+}
+
+impl DatabaseTransaction {
+    /// Commit the transaction
+    pub async fn commit(self) -> Result<(), DatabaseError> {
+        println!("Committing transaction on connection {}", self.connection_id);
+        Ok(())
+    }
+
+    /// Rollback the transaction
+    pub async fn rollback(self) -> Result<(), DatabaseError> {
+        println!("Rolling back transaction on connection {}", self.connection_id);
+        Ok(())
+    }
+}
+
+/// Database error types
+#[derive(Debug)]
+pub enum DatabaseError {
+    ConnectionFailed,
+    QueryFailed(String),
+    PoolExhausted,
+    TransactionFailed,
+}
+
+impl std::fmt::Display for DatabaseError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            DatabaseError::ConnectionFailed => write!(f, "Failed to connect to database"),
+            DatabaseError::QueryFailed(msg) => write!(f, "Query failed: {}", msg),
+            DatabaseError::PoolExhausted => write!(f, "Database connection pool exhausted"),
+            DatabaseError::TransactionFailed => write!(f, "Transaction failed"),
+        }
+    }
+}
+
+impl std::error::Error for DatabaseError {}
+
+/// Connect to the database using the provided URL
+pub fn connect(database_url: &str) -> Result<(), DatabaseError> {
+    println!("Connecting to database: {}", database_url);
+    // Mock implementation
+    Ok(())
+}
